Separate dev harnesses for CompatibilityAnalsysis and ModelComparison
