--- .flaskenv
@@ -1,3 +1,0 @@
-FLASK_APP=development/app.py
-FLASK_ENV=development
-FLASK_RUN_PORT=5000
