+++ development/model-comparison/css/custom.css
@@ -1,0 +1,8 @@
+/*
+Placeholder for custom user CSS
+
+mainly to be overridden in profile/static/custom/custom.css
+
+This will always be an empty file
+*/
+
