+++ package.json
@@ -53,7 +53,8 @@
     "webpack-dev-server": "^3.11.0"
   },
   "scripts": {
-    "start": "webpack-dev-server --hot --mode development --host 0.0.0.0",
+    "start-compatibility-analysis": "webpack-dev-server --config ./compatibilityanalysis.webpack.config.js --hot --mode development --host 0.0.0.0",
+    "start-model-comparison": "webpack-dev-server --config ./modelcomparison.webpack.config.js --hot --mode development --host 0.0.0.0",
     "build": "cross-env NODE_ENV=production webpack --config ./compatibilityanalysis.webpack.config.js && cross-env NODE_ENV=production webpack --config ./modelcomparison.webpack.config.js",
     "test": "jest"
   }
